aam_gauntlet/api_farm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6